Cleaning code.
