--- src/player/duplexStream.js
@@ -1,7 +1,0 @@
-let Duplex = require("stream").Duplex;
-
-module.exports = class DuplexStream {
-    constructor() {
-        let d = new Duplex();
-    }
-}
