+++ src/player/track.js
@@ -1,5 +1,3 @@
-//import { Readable } from "stream";
-
 "use strict";
 
 module.exports = class Track {
